Additional refactoring
* Unit Test merged into TestGracefulShutdown class * Leveraged reflections to get handle on WebServer's temp directory
* Removed unnecessary File delete calls
